camera: move Camera_Fixed to TRX
